requirements.txt
anthropic>=0.45.1
httpx>=0.28.1
mcp[cli]>=1.2.1
python-dotenv>=1.0.1
argparse>=1.4.0
boto3>=1.33.0
